agent/pyproject.toml
[tool.poetry]
name = "agent"
version = "0.1.0"
description = "LiveKit Audio Agent"
authors = ["songdy <you@example.com>"]
packages = [{ include = "livekit" }]

[tool.poetry.dependencies]
python = "^3.10"
livekit = "*"
aiohttp = "*"
boto3 = "*"
openai = "*"
# 필요한 패키지 더 추가 가능
openai-whisper = "^20250625"
av = "^15.0.0"
livekit-agents = "^1.2.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
